pyproject.toml
[tool.poetry]
name = "openfeature-provider-python"
version = "1.0.0"
description = "Hyphen provider for OpenFeature Python SDK"
authors = ["Hyphen AI"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/Hyphen/openfeature-provider-python"
packages = [
    { include = "openfeature_provider_hyphen", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.8"
openfeature-sdk = "0.7.4"
requests = "^2.31.0"
cachetools = "^5.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
requests-mock = "^1.11.0"
isort = "^5.12.0"
ruff = "^0.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = ["pyproject.toml:version"]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
version_source = "commit"
commit_version_number = true
branch = "main"
upload_to_pypi = true
upload_to_release = true
build_command = "pip install poetry && poetry build"
dist_path = "dist/"
remove_dist = false
patch_without_tag = true

